+++ custom.js
@@ -1,8 +1,10 @@
 'use strict'
 
-// Deprecated export, use the `.default` export instead.
-exports = module.exports = require('./commonjs/Input').default
+exports = module.exports = require('./commonjs/PhoneInputReactResponsiveUI').default
 
+exports.PhoneInputNative = require('./commonjs/PhoneInputNative').default
+
+// Deprecated export, use `smartCaret={false}` instead of `inputComponent={BasicInput}`.
 exports.BasicInput = require('./commonjs/BasicInput').default
 
-exports['default'] = require('./commonjs/Input').default
+exports['default'] = require('./commonjs/PhoneInputReactResponsiveUI').default