+++ index.js
@@ -1,4 +1,9 @@
+export { default as PhoneInputNative } from './modules/PhoneInputNativeDefaultMetadata'
+export { default as PhoneInputReactResponsiveUI } from './modules/PhoneInputReactResponsiveUIDefaultMetadata'
+export { default as default } from './modules/PhoneInputReactResponsiveUIDefaultMetadata'
+
+// Deprecated export. Will be removed in some future major version.
+export { default as Input } from './modules/PhoneInputReactResponsiveUI'
+
 // Deprecated export, use `smartCaret={false}` instead of `inputComponent={BasicInput}`.
 export { default as BasicInput } from './modules/BasicInput'
-export { default as Input } from './modules/Input'
-export { default as default } from './modules/InputWithDefaultMetadata'