+++ source/InputWithDefaultMetadata.js
@@ -1,22 +1,2 @@
-import React, { Component } from 'react'
-import metadata from 'libphonenumber-js/metadata.min'
-
-import Input from './Input'
-
-export default class InputWithDefaultMetadata extends Component
-{
-	render()
-	{
-		return (
-			<Input
-				ref={ ref => this.input = ref }
-				{ ...this.props }
-				metadata={ metadata }/>
-		)
-	}
-
-	focus()
-	{
-		return this.input.focus()
-	}
-}
+// Deprecated file. Will be removed in some future major version.
+export { default as default } from './PhoneInputReactResponsiveUIDefaultMetadata'