+++ test/exports.test.js
@@ -1,5 +1,6 @@
-import Phone,
+import PhoneInput,
 {
+	PhoneInputNative,
 	Input,
 	BasicInput
 }
@@ -9,7 +10,9 @@ describe(`exports`, function()
 {
 	it(`should export ES6`, function()
 	{
-		Phone.should.be.a('function')
+		PhoneInput.should.be.a('function')
+		PhoneInputNative.should.be.a('function')
+
 		Input.should.be.a('function')
 		BasicInput.should.be.a('function')
 	})
@@ -23,9 +26,13 @@ describe(`exports`, function()
 		// Deprecated export.
 		Library.should.be.a('function')
 
+		Library.PhoneInputNative.should.be.a('function')
+
 		Custom.default.should.be.a('function')
 		// Deprecated export.
 		Custom.should.be.a('function')
+
+		Custom.PhoneInputNative.should.be.a('function')
 
 		// Duplicating this export.
 		Library.BasicInput.should.be.a('function')